+++ packages/umi-uploader-arweave-via-turbo/src/index.ts
@@ -1,3 +1,3 @@
-export * from './createBundlrUploader';
+export * from './createArweaveUploader';
 export * from './errors';
 export * from './plugin';