+++ packages/umi-uploader-irys/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @metaplex-foundation/umi-uploader-irys
 
+## 1.0.0
+
+### Major Changes
+
+- [`d50553a`](https://github.com/metaplex-foundation/umi/commit/d50553a419b7c9beac996a7c0b4d5942c91b5b4e) Thanks [@blockiosaurus](https://github.com/blockiosaurus)! - Release 1.0
+
+### Minor Changes
+
+- [#156](https://github.com/metaplex-foundation/umi/pull/156) [`c33df61`](https://github.com/metaplex-foundation/umi/commit/c33df615681f56f502c496c59d7ab418dff24c9a) Thanks [@JesseTheRobot](https://github.com/JesseTheRobot)! - Update uploader-irys to use the latest packages
+
+### Patch Changes
+
+- [#145](https://github.com/metaplex-foundation/umi/pull/145) [`48bb4fd`](https://github.com/metaplex-foundation/umi/commit/48bb4fdb92cb4e131e301628796df4b6af982b89) Thanks [@blockiosaurus](https://github.com/blockiosaurus)! - Updating irys to latest
+
+- [#148](https://github.com/metaplex-foundation/umi/pull/148) [`597fcad`](https://github.com/metaplex-foundation/umi/commit/597fcad8a309f770fdd302f3771cc934345f788a) Thanks [@rocketfren](https://github.com/rocketfren)! - Update upload to return Irys gateway URLs to Arweave instead of the arweave.net Gateway. This fixes current devnet issues.
+
+- Updated dependencies [[`48bb4fd`](https://github.com/metaplex-foundation/umi/commit/48bb4fdb92cb4e131e301628796df4b6af982b89), [`7404903`](https://github.com/metaplex-foundation/umi/commit/7404903e58fe519c7d79b7c0be5389cb16398fb7), [`e192236`](https://github.com/metaplex-foundation/umi/commit/e1922366470201bb51639a3a2660fee3a4cc7f17), [`d50553a`](https://github.com/metaplex-foundation/umi/commit/d50553a419b7c9beac996a7c0b4d5942c91b5b4e), [`91a4d75`](https://github.com/metaplex-foundation/umi/commit/91a4d75484ce6d65e30e29646539c18d88ee1f80), [`8c8aa5f`](https://github.com/metaplex-foundation/umi/commit/8c8aa5f4f2dc22872870cb4824c3672d6f8ac0ca)]:
+  - @metaplex-foundation/umi@1.0.0
+  - @metaplex-foundation/umi-web3js-adapters@1.0.0
+
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies [[`6276f96`](https://github.com/metaplex-foundation/umi/commit/6276f9651f91b973b474cb63a0457917b09ed292)]:
+  - @metaplex-foundation/umi@0.9.2
+  - @metaplex-foundation/umi-web3js-adapters@0.9.2
+
 ## 0.9.1
 
 ### Patch Changes